AB#100570 move address to types folder
