+++ packages/forms/src/elements/address/addressLookupProvider.ts
@@ -1,4 +1,4 @@
-import { Address } from './address';
+import { Address } from './types';
 
 export interface AddressLookupProvider {
 	lookupAddress(postcode: string, limit?: number): Promise<any>;
