+++ packages/forms/src/elements/address/types/index.ts
@@ -2,3 +2,4 @@ export * from './AddressProps';
 export * from './BaseProps';
 export * from './SelectAddressProps';
 export * from './EditAddressProps';
+export * from './Address';
